requirements.txt
Flask==3.1.3
google-genai==1.74.0
python-dotenv==1.2.2
